Move estimateGas related logic to separate file
